Add first version of animation state machinery
Add first version of animation state machinery. There's a new class
AnimatorStateClass that defines an entity class specific animation
state machine, i.e. states and their transitions. Each entity will
then have an animator instance of this type. The animator can have
an associated Lua script which will be invoked on animator state
changes in order to evaluate state exit conditions and do actions
on state leave, enter and update.

+++ editor/gui/animationtrackwidget.h
@@ -186,11 +186,11 @@ namespace gui
         } mState;
         bool mCameraWasLoaded = false;
         // Time accumulator for animation playback only.
-        float mAnimationTime = 0.0f;
+        double mAnimationTime = 0.0f;
         // Current time accumulator used by the widget itself.
-        float mCurrentTime = 0.0f;
+        double mCurrentTime = 0.0f;
         // Interpolation variables for smoothing view rotations.
-        float mViewTransformStartTime = 0.0f;
+        double mViewTransformStartTime = 0.0f;
         float mViewTransformRotation = 0.0f;
         // Entity object is used render the animation track
         // while it's being edited. We use an entity object
